Move deprecated methods & properties into legacy config class
