+++ tests/Unit/Config/GeneralConfigTest.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use craft\config\GeneralConfig as LegacyGeneralConfig;
 use CraftCms\Cms\Cms;
 use CraftCms\Cms\Config\ConfigServiceProvider;
 use CraftCms\Cms\Config\GeneralConfig;
@@ -59,4 +60,24 @@ it('can set trackedQueueNames via fluent setter', function () {
     $config = GeneralConfig::create()->trackedQueueNames(['craft', 'default']);
 
     expect($config->trackedQueueNames)->toBe(['craft', 'default']);
+});
+
+it('does not expose moved deprecated members on the new class', function () {
+    $config = GeneralConfig::create();
+
+    expect(method_exists($config, 'devMode'))->toBeFalse()
+        ->and(method_exists($config, 'enableCsrfProtection'))->toBeFalse()
+        ->and(method_exists($config, 'securityKey'))->toBeFalse()
+        ->and(property_exists($config, 'allowedGraphqlOrigins'))->toBeFalse()
+        ->and(property_exists($config, 'userSessionDuration'))->toBeFalse();
+});
+
+it('keeps moved deprecated members on the legacy class', function () {
+    $config = LegacyGeneralConfig::create();
+
+    expect(method_exists($config, 'devMode'))->toBeTrue()
+        ->and(method_exists($config, 'enableCsrfProtection'))->toBeTrue()
+        ->and(method_exists($config, 'securityKey'))->toBeTrue()
+        ->and(property_exists($config, 'allowedGraphqlOrigins'))->toBeTrue()
+        ->and(property_exists($config, 'userSessionDuration'))->toBeTrue();
 });
